+++ apps/sim/lib/webhooks/deploy.ts
@@ -1,11 +1,5 @@
 import { db } from '@sim/db'
-import {
-  account,
-  credentialSetMember,
-  webhook,
-  workflow,
-  workflowDeploymentVersion,
-} from '@sim/db/schema'
+import { account, credentialSetMember, webhook, workflowDeploymentVersion } from '@sim/db/schema'
 import { createLogger } from '@sim/logger'
 import { generateShortId } from '@sim/utils/id'
 import { and, eq, inArray, isNotNull, isNull, or } from 'drizzle-orm'
@@ -18,7 +12,10 @@ import {
   shouldRecreateExternalWebhookSubscription,
 } from '@/lib/webhooks/provider-subscriptions'
 import { getProviderHandler } from '@/lib/webhooks/providers'
-import { syncWebhooksForCredentialSet } from '@/lib/webhooks/utils.server'
+import {
+  findConflictingWebhookPathOwner,
+  syncWebhooksForCredentialSet,
+} from '@/lib/webhooks/utils.server'
 import { buildCanonicalIndex } from '@/lib/workflows/subblocks/visibility'
 import { getBlock } from '@/blocks'
 import type { SubBlockConfig } from '@/blocks/types'
@@ -28,37 +25,6 @@ import { SYSTEM_SUBBLOCK_IDS } from '@/triggers/constants'
 
 const logger = createLogger('DeployWebhookSync')
 const CREDENTIAL_SET_PREFIX = 'credentialSet:'
-
-/**
- * Returns the id of a different workflow that already owns an active webhook on
- * the given path, or `null` if the path is free or owned by this workflow.
- * Guards against cross-tenant path collisions, since paths are user-controlled
- * and only unique per deployment version. Mirrors the runtime dispatcher's
- * filter (active, non-archived webhook on a non-archived workflow) so webhooks
- * that can never receive deliveries don't permanently reserve a path.
- */
-async function findConflictingWebhookPathOwner(params: {
-  path: string
-  workflowId: string
-}): Promise<string | null> {
-  const { path, workflowId } = params
-
-  const existing = await db
-    .select({ workflowId: webhook.workflowId })
-    .from(webhook)
-    .innerJoin(workflow, eq(webhook.workflowId, workflow.id))
-    .where(
-      and(
-        eq(webhook.path, path),
-        eq(webhook.isActive, true),
-        isNull(webhook.archivedAt),
-        isNull(workflow.archivedAt)
-      )
-    )
-
-  const conflict = existing.find((row) => row.workflowId !== workflowId)
-  return conflict ? conflict.workflowId : null
-}
 
 interface TriggerSaveError {
   message: string
